# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stepflow API",
    "description": "API for Stepflow flows and runs",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.9.0"
  },
  "servers": [
    {
      "url": "http://localhost:7840/api/v1",
      "description": "Default development server (port 7840)"
    }
  ],
  "paths": {
...
    "/blobs": {
      "post": {
        "tags": [
          "Blob"
        ],
        "summary": "Store a blob",
        "description": "Store a blob and return its content-based ID. Supports two content types: `application/json` (JSON body with `data`, `blobType`, and optional `filename` fields) and `application/octet-stream` (raw binary body, use `X-Blob-Filename` header for filename).",
        "operationId": "storeBlob",
        "requestBody": {
          "content": {
            "application/octet-stream": {}
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Blob stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreBlobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/blobs/{blob_id}": {
      "get": {
        "tags": [
          "Blob"
        ],
        "summary": "Get a blob by ID",
        "description": "Retrieve a blob by its content-based ID. Supports content negotiation via `Accept` header: `application/json` (default) returns JSON with data, blobType, blobId, and filename; `application/octet-stream` returns raw bytes.",
        "operationId": "getBlob",
        "responses": {
          "200": {
            "description": "Blob data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBlobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Blob not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "blob_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/flows": {
      "post": {
        "tags": [
          "Flow"
        ],
        "summary": "Store a flow",
        "description": "Store a flow and return its content-based hash ID.",
        "operationId": "storeFlow",
        "requestBody": {
          "description": "Request to store a flow",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreFlowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response when a flow is stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreFlowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid flow definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/flows/{flow_id}": {
      "get": {
        "tags": [
          "Flow"
        ],
        "summary": "Get a flow by ID",
        "description": "Retrieve a flow definition by its content-based hash ID.",
        "operationId": "getFlow",
        "responses": {
          "200": {
            "description": "Response containing a flow definition and its ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResponse"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Flow"
        ],
        "summary": "Delete a flow by ID",
        "description": "Delete a flow by its content-based hash ID.",
        "operationId": "deleteFlow",
        "responses": {
          "200": {
            "description": "no content"
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "FlowResponse": {
        "description": "Response containing a flow definition and its ID",
        "type": "object",
        "properties": {
          "flow": {
            "description": "The flow definition",
            "allOf": [
              {
                "$ref": "#/components/schemas/Flow"
              }
            ]
          },
          "flowId": {
            "description": "The flow ID",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlobId"
              }
            ]
          },
          "allExamples": {
            "description": "All available examples (includes both examples and test cases)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExampleInput"
            }
          }
        },
        "required": [
          "flow",
          "flowId",
          "allExamples"
        ]
      },
...
      "ErrorResponse": {
        "description": "Error response structure.\n\nServer handlers should return this, but usually it is better to create it\nby returning an `error_stack::Report<ServerError>` and using the automatic\nconversion to `ErrorResponse`.\n\nOther `error_stack::Report` types will automatically convert to internal errors.",
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code (e.g., 400, 404, 500)",
            "type": "integer",
            "format": "uint16",
            "minimum": 0,
            "maximum": 65535
          },
          "message": {
            "description": "Human-readable error message",
            "type": "string"
          },
          "stack": {
            "description": "Detailed error stack for debugging",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorStackEntry"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorStackEntry": {
        "description": "A single entry in an error stack for detailed error reporting",
        "type": "object",
        "properties": {
          "error": {
            "description": "The error message",
            "type": "string"
          },
          "attachments": {
            "description": "Additional context attached to this error",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "backtrace": {
            "description": "Backtrace if available",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "StoreBlobResponse": {
        "description": "Response when a blob is stored",
        "type": "object",
        "properties": {
          "blobId": {
            "description": "The content-based blob ID (SHA-256 hash)",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlobId"
              }
            ]
          },
          "filename": {
            "description": "The filename if one was provided",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "blobId"
        ]
      },
      "BlobId": {
        "description": "A SHA-256 hash of the blob content, represented as a hexadecimal string.",
        "type": "string"
      },
      "GetBlobResponse": {
        "description": "Response when retrieving a blob (JSON mode)",
        "type": "object",
        "properties": {
          "data": {
            "description": "The blob data"
          },
          "blobType": {
            "description": "The blob type",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlobType"
              }
            ]
          },
          "blobId": {
            "description": "The blob ID (for confirmation)",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlobId"
              }
            ]
          },
          "filename": {
            "description": "The filename if one was set",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data",
          "blobType",
          "blobId"
        ]
      },
      "BlobType": {
        "description": "Type of blob stored in the blob store.",
        "oneOf": [
          {
            "description": "A workflow/flow definition",
            "type": "string",
            "const": "flow"
          },
          {
            "description": "Generic data blob",
            "type": "string",
            "const": "data"
          },
          {
            "description": "Raw binary data (base64-encoded in JSON transport)",
            "type": "string",
            "const": "binary"
          }
        ]
      },
...
      "Schema": {
        "description": "A valid JSON Schema object."
      },
...
      "Value": {
        "description": "Any JSON value (object, array, string, number, boolean, or null)"
      },
...
      "Flow": {
        "description": "A workflow consisting of a sequence of steps and their outputs.\n\nA flow represents a complete workflow that can be executed. It contains:\n- A sequence of steps to execute\n- Named outputs that can reference step outputs\n\nFlows should not be cloned. They should generally be stored and passed as a\nreference or inside an `Arc`.",
        "type": "object",
        "properties": {
          "name": {
            "description": "The name of the flow.",
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "description": "The description of the flow.",
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "description": "The version of the flow.",
            "type": [
              "string",
              "null"
            ]
          },
          "schemas": {
            "description": "Consolidated schema information for the flow.\nContains input/output schemas, step output schemas, and shared `$defs`.",
            "allOf": [
              {
                "$ref": "#/components/schemas/FlowSchema"
              }
            ]
          },
          "steps": {
            "description": "The steps to execute for the flow.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Step"
            }
          },
          "output": {
            "description": "The outputs of the flow, mapping output names to their values.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ValueExpr"
              }
            ]
          },
          "test": {
            "description": "Test configuration for the flow.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "examples": {
            "description": "Example inputs for the workflow that can be used for testing and UI dropdowns.",
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ExampleInput"
            }
          },
          "metadata": {
            "description": "Extensible metadata for the flow that can be used by tools and frameworks.",
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "FlowSchema": {
        "description": "Consolidated schema information for a flow.\n\nThis struct contains all schema/type information for the flow in a single location,\nallowing shared `$defs` across all schemas and avoiding duplication.\n\nSerializes as a valid JSON Schema with `type: \"object\"` and flow-specific\nproperties (`input`, `output`, `variables`, `steps`) under the `properties` key.",
        "type": "object",
        "properties": {
          "defs": {
            "description": "Shared type definitions that can be referenced by other schemas.\nReferences use the format `#/schemas/$defs/TypeName`.",
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Schema"
            }
          },
          "input": {
            "description": "The input schema for the flow.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "type": "null"
              }
            ]
          },
          "output": {
            "description": "The output schema for the flow.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "type": "null"
              }
            ]
          },
          "variables": {
            "description": "Schema for workflow variables. This is a JSON Schema object where\nproperties define the available variables and their types.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "type": "null"
              }
            ]
          },
          "steps": {
            "description": "Output schemas for each step, keyed by step ID.\nNote: Step input schemas are not included here as they are\ncomponent metadata, not flow-specific schemas.\nUses IndexMap to preserve insertion order for deterministic serialization.",
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Schema"
            }
          }
        },
        "required": [
          "defs",
          "steps"
        ]
      },
...
      "TestConfig": {
        "description": "Configuration for testing a workflow.",
        "type": "object",
        "properties": {
          "configFile": {
            "description": "Path to an external stepflow config file for tests.\nRelative paths are resolved from the workflow file's directory.\nMutually exclusive with `config` - validated at runtime.",
            "type": [
              "string",
              "null"
            ]
          },
          "config": {
            "description": "Inline stepflow configuration for tests.\nMutually exclusive with `config_file` - validated at runtime."
          },
          "cases": {
            "description": "Test cases for the workflow.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestCase"
            }
          }
        }
      },
...
      "ExampleInput": {
        "description": "An example input for a workflow that can be used in UI dropdowns.",
        "type": "object",
        "properties": {
          "name": {
            "description": "Name of the example input for display purposes.",
            "type": "string"
          },
          "description": {
            "description": "Optional description of what this example demonstrates.",
            "type": [
              "string",
              "null"
            ]
          },
          "input": {
            "description": "The input data for this example.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Value"
              }
            ]
          }
        },
        "required": [
          "name",
          "input"
        ]
      },
      "Diagnostics": {
        "description": "Collection of diagnostics with utility methods",
        "type": "object",
        "properties": {
          "diagnostics": {
            "description": "All diagnostics found",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            }
          },
          "numFatal": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "numError": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "numWarning": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          }
        },
        "required": [
          "diagnostics",
          "numFatal",
          "numError",
          "numWarning"
        ]
      },
...
      "StoreFlowRequest": {
        "description": "Request to store a flow",
        "type": "object",
        "properties": {
          "flow": {
            "description": "The flow to store",
            "allOf": [
              {
                "$ref": "#/components/schemas/Flow"
              }
            ]
          },
          "dryRun": {
            "description": "If true, only validate the flow without storing it",
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "flow"
        ]
      },
      "StoreFlowResponse": {
        "description": "Response when a flow is stored",
        "type": "object",
        "properties": {
          "flowId": {
            "description": "The ID of the flow (computed from content hash, always present)",
            "allOf": [
              {
                "$ref": "#/components/schemas/BlobId"
              }
            ]
          },
          "stored": {
            "description": "Whether the flow was actually stored (false for dry_run or if validation has fatal errors)",
            "type": "boolean"
          },
          "diagnostics": {
            "description": "Validation diagnostics",
            "allOf": [
              {
                "$ref": "#/components/schemas/Diagnostics"
              }
            ]
          }
        },
        "required": [
          "flowId",
          "stored",
          "diagnostics"
        ]
      }
    }
  }
}